# multas-backend/docs/API_OPENAPI.yaml
openapi: 3.1.0
info:
  title: PMT Multas API - TANDAS 0–8, 10 y 11
  version: 1.2.0
  description: Contrato de autenticación, operación municipal, finanzas, solvencias, analítica, notificaciones y migración histórica controlada.
servers:
  - url: /api/v1
...
paths:
...
  /citizens:
    get:
      tags: [Citizens]
      security: [{ sessionCookie: [] }]
      summary: Lista ciudadanos con búsqueda y paginación
      responses:
        '200': { description: Página de ciudadanos; datos sensibles según permiso }
    post:
      tags: [Citizens]
      security: [{ sessionCookie: [] }]
      summary: Crea un ciudadano
      responses:
        '201': { description: Ciudadano creado }
        '409': { $ref: '#/components/responses/Error' }
  /citizens/{id}:
    get:
      tags: [Citizens]
      security: [{ sessionCookie: [] }]
      summary: Consulta un ciudadano
      responses:
        '200': { description: Ciudadano }
    patch:
      tags: [Citizens]
      security: [{ sessionCookie: [] }]
      summary: Actualiza un ciudadano
      responses:
        '204': { description: Ciudadano actualizado }
...
  /devices:
    get:
      tags: [Administration]
      security: [{ sessionCookie: [] }]
      summary: Lista dispositivos y asignación vigente
      responses: { '200': { description: Página de dispositivos } }
    post:
      tags: [Administration]
      security: [{ sessionCookie: [] }]
      summary: Registra un dispositivo institucional
      responses: { '201': { description: Dispositivo creado }, '409': { $ref: '#/components/responses/Error' } }
  /devices/{id}/assign:
    post:
      tags: [Administration]
      security: [{ sessionCookie: [] }]
      summary: Asigna un dispositivo conservando historial
      responses: { '201': { description: Asignación creada }, '409': { $ref: '#/components/responses/Error' } }
  /devices/{id}:
    get:
      tags: [Administration]
      security: [{ sessionCookie: [] }]
      summary: Consulta dispositivo, asignación vigente e historial
      responses: { '200': { description: Dispositivo institucional }, '404': { $ref: '#/components/responses/Error' } }
    patch:
      tags: [Administration]
      security: [{ sessionCookie: [] }]
      summary: Actualiza datos del dispositivo
      responses: { '204': { description: Dispositivo actualizado }, '409': { $ref: '#/components/responses/Error' } }
  /devices/{id}/unassign:
    post:
      tags: [Administration]
      security: [{ sessionCookie: [] }]
      summary: Libera la asignación vigente y conserva el historial
      responses: { '204': { description: Dispositivo liberado }, '404': { $ref: '#/components/responses/Error' } }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pmt_session
  responses:
    Error:
      description: Error estable con requestId
      content:
        application/json:
          schema:
            type: object
            required: [error, meta]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
              meta:
                type: object
                required: [requestId]
                properties:
                  requestId: { type: string, format: uuid }