# multas-backend/src/modules/devices/http/device.routes.ts
import { Router } from "express";
import type { ResultSetHeader, RowDataPacket } from "mysql2/promise";
import { z } from "zod";
import type { AppContainer } from "../../../bootstrap/container.js";
import { authenticate } from "../../../shared/http/authenticate.js";
import { authorize } from "../../../shared/http/authorize.js";
import { HttpError } from "../../../shared/http/HttpError.js";
import { getRequestId } from "../../../shared/http/request-context.js";
import { isDuplicateKey, recordOperation } from "../../../shared/http/operations.js";

const deviceInput=z.object({deviceUuid:z.string().trim().min(8).max(100),institutionalCode:z.string().trim().min(2).max(50),deviceType:z.enum(["MOBILE","WEB_STATION"]),platform:z.string().trim().min(2).max(30),model:z.string().trim().max(100).nullable().optional(),operatingSystem:z.string().trim().max(100).nullable().optional(),appVersion:z.string().trim().max(30).nullable().optional()});
const deviceUpdate=deviceInput.partial().refine(value=>Object.keys(value).length>0);
const assignment=z.object({userId:z.coerce.number().int().positive()});
const release=z.object({reason:z.string().trim().min(3).max(300)});
const list=z.object({search:z.string().trim().max(100).default(""),status:z.enum(["PENDING","ACTIVE","BLOCKED","RETIRED"]).optional(),page:z.coerce.number().int().min(1).default(1),pageSize:z.coerce.number().int().min(1).max(100).default(20)});

export function createDeviceRouter(container:AppContainer):Router{
 const router=Router();router.use(authenticate(container.authRepository,container.env));
 router.get("/",authorize("devices.read",container.auditRepository),async(request,response,next)=>{try{const q=list.parse(request.query);const clauses:string[]=[];const values:unknown[]=[];if(q.status){clauses.push("d.status=?");values.push(q.status);}if(q.search){clauses.push("(d.institutional_code LIKE ? OR d.device_uuid LIKE ? OR d.model LIKE ?)");const like=`%${q.search}%`;values.push(like,like,like);}const where=clauses.length?`WHERE ${clauses.join(" AND ")}`:"";const counts=await container.database.query<(RowDataPacket&{total:number})[]>(`SELECT COUNT(*) total FROM devices d ${where}`,values);const rows=await container.database.query<RowDataPacket[]>(`SELECT d.id,d.device_uuid,d.institutional_code,d.device_type,d.platform,d.model,d.operating_system,d.app_version,d.status,d.last_seen_at,d.created_at,da.id assignment_id,da.user_id,u.username,u.first_name,u.last_name FROM devices d LEFT JOIN device_assignments da ON da.device_id=d.id AND da.released_at IS NULL LEFT JOIN users u ON u.id=da.user_id ${where} ORDER BY d.institutional_code LIMIT ? OFFSET ?`,[...values,q.pageSize,(q.page-1)*q.pageSize]);response.json({data:rows,meta:{page:q.page,pageSize:q.pageSize,total:counts[0]?.total??0,requestId:getRequestId()}});}catch(error){next(error);}});
 router.get("/:id",authorize("devices.read",container.auditRepository),async(request,response,next)=>{try{const id=z.coerce.number().int().positive().parse(request.params["id"]);const rows=await container.database.query<RowDataPacket[]>("SELECT d.*,INET6_NTOA(d.last_ip) last_ip,da.id assignment_id,da.user_id,u.username,u.first_name,u.last_name FROM devices d LEFT JOIN device_assignments da ON da.device_id=d.id AND da.released_at IS NULL LEFT JOIN users u ON u.id=da.user_id WHERE d.id=?",[id]);if(!rows[0])throw new HttpError({code:"DEVICE_NOT_FOUND",message:"Dispositivo no encontrado.",statusCode:404});const history=await container.database.query<RowDataPacket[]>("SELECT da.id,da.user_id,da.assigned_at,da.released_at,da.release_reason,u.username,u.first_name,u.last_name FROM device_assignments da JOIN users u ON u.id=da.user_id WHERE da.device_id=? ORDER BY da.assigned_at DESC",[id]);response.json({data:{...rows[0],assignments:history},meta:{requestId:getRequestId()}});}catch(error){next(error);}});
 router.post("/",authorize("devices.manage",container.auditRepository),async(request,response,next)=>{try{const input=deviceInput.parse(request.body);const result=await container.database.query<ResultSetHeader>("INSERT INTO devices (device_uuid,institutional_code,device_type,platform,model,operating_system,app_version,status) VALUES (?,?,?,?,?,?,?,'PENDING')",[input.deviceUuid,input.institutionalCode,input.deviceType,input.platform,input.model??null,input.operatingSystem??null,input.appVersion??null]);await recordOperation(container,request,{action:"DEVICE_CREATED",module:"devices",entityType:"device",entityId:String(result.insertId)});response.status(201).json({data:{id:String(result.insertId)},meta:{requestId:getRequestId()}});}catch(error){if(isDuplicateKey(error))return next(new HttpError({code:"DEVICE_DUPLICATE",message:"El identificador del dispositivo ya existe.",statusCode:409}));next(error);}});
 router.patch("/:id",authorize("devices.manage",container.auditRepository),async(request,response,next)=>{try{const id=z.coerce.number().int().positive().parse(request.params["id"]);const input=deviceUpdate.parse(request.body);const cols:string[]=[];const vals:unknown[]=[];const map={deviceUuid:"device_uuid",institutionalCode:"institutional_code",deviceType:"device_type",platform:"platform",model:"model",operatingSystem:"operating_system",appVersion:"app_version"} as const;for(const [key,col] of Object.entries(map) as [keyof typeof map,string][])if(input[key]!==undefined){cols.push(`${col}=?`);vals.push(input[key]);}const result=await container.database.query<ResultSetHeader>(`UPDATE devices SET ${cols.join(",")} WHERE id=?`,[...vals,id]);if(!result.affectedRows)throw new HttpError({code:"DEVICE_NOT_FOUND",message:"Dispositivo no encontrado.",statusCode:404});await recordOperation(container,request,{action:"DEVICE_UPDATED",module:"devices",entityType:"device",entityId:String(id)});response.sendStatus(204);}catch(error){if(isDuplicateKey(error))return next(new HttpError({code:"DEVICE_DUPLICATE",message:"El identificador del dispositivo ya existe.",statusCode:409}));next(error);}});
 router.post("/:id/assign",authorize("devices.assign",container.auditRepository),async(request,response,next)=>{try{const id=z.coerce.number().int().positive().parse(request.params["id"]);const input=assignment.parse(request.body);const assignmentId=await container.database.withTransaction(async(connection)=>{const [devices]=await connection.query<(RowDataPacket&{status:string})[]>("SELECT status FROM devices WHERE id=? FOR UPDATE",[id]);if(!devices[0])throw new HttpError({code:"DEVICE_NOT_FOUND",message:"Dispositivo no encontrado.",statusCode:404});if(devices[0].status==="BLOCKED"||devices[0].status==="RETIRED")throw new HttpError({code:"DEVICE_UNAVAILABLE",message:"El dispositivo bloqueado o retirado no puede asignarse.",statusCode:409});const [users]=await connection.query<(RowDataPacket&{status:string})[]>("SELECT status FROM users WHERE id=?",[input.userId]);if(users[0]?.status!=="ACTIVE")throw new HttpError({code:"USER_INACTIVE",message:"El responsable debe ser un usuario activo.",statusCode:409});const [result]=await connection.query<ResultSetHeader>("INSERT INTO device_assignments (device_id,user_id,assigned_by_user_id) VALUES (?,?,?)",[id,input.userId,request.auth?.user.id??null]);await connection.query("UPDATE devices SET status='ACTIVE' WHERE id=?",[id]);return String(result.insertId);});await recordOperation(container,request,{action:"DEVICE_ASSIGNED",module:"devices",entityType:"device_assignment",entityId:assignmentId});response.status(201).json({data:{id:assignmentId},meta:{requestId:getRequestId()}});}catch(error){if(isDuplicateKey(error))return next(new HttpError({code:"DEVICE_ALREADY_ASSIGNED",message:"El dispositivo ya tiene una asignación activa.",statusCode:409}));next(error);}});
 router.post("/:id/unassign",authorize("devices.assign",container.auditRepository),async(request,response,next)=>{try{const id=z.coerce.number().int().positive().parse(request.params["id"]);const {reason}=release.parse(request.body);const result=await container.database.query<ResultSetHeader>("UPDATE device_assignments SET released_at=UTC_TIMESTAMP(3),release_reason=? WHERE device_id=? AND released_at IS NULL",[reason,id]);if(!result.affectedRows)throw new HttpError({code:"DEVICE_NOT_ASSIGNED",message:"No existe una asignación activa.",statusCode:404});await recordOperation(container,request,{action:"DEVICE_UNASSIGNED",module:"devices",entityType:"device",entityId:String(id),reason});response.sendStatus(204);}catch(error){next(error);}});
 for(const [path,status,action] of [["block","BLOCKED","DEVICE_BLOCKED"],["unblock","ACTIVE","DEVICE_UNBLOCKED"]] as const)router.post(`/:id/${path}`,authorize("devices.manage",container.auditRepository),async(request,response,next)=>{try{const id=z.coerce.number().int().positive().parse(request.params["id"]);const result=await container.database.query<ResultSetHeader>("UPDATE devices SET status=? WHERE id=? AND status<>'RETIRED'",[status,id]);if(!result.affectedRows)throw new HttpError({code:"DEVICE_NOT_FOUND",message:"Dispositivo no encontrado o retirado.",statusCode:404});await recordOperation(container,request,{action,module:"devices",entityType:"device",entityId:String(id)});response.sendStatus(204);}catch(error){next(error);}});
 return router;
}